requirements.txt
kafka-python
gssapi
requests
pymysql
uuid
flask
flask_login
flask_admin
flask_sqlalchemy